Create an api service
