+++ frontend/app/mock-api.ts
@@ -1,0 +1,3 @@
+import { GameInfo } from './models';
+
+export const GAMES: GameInfo[] = [];
